Update 3: Map Window Centering
-Map window now centers on right clicking
-Renamed Vec3f to Vec3f for consistency
-Added mouse clicking callbacks / handling code
-Removed screenToWorld from helpers (since we have a local scrToMap for
the map and everything else is 1:1)
-Added text functions to helpers
-Next goal: Map zooming

+++ General/Vec3d.cpp
@@ -1,12 +1,12 @@
-#include "Vec3f.h"
+#include "Vec3d.h"
 
-Vec3f::Vec3f() {
+Vec3d::Vec3d() {
     this->x = 0.0;
     this->y = 0.0;
     this->z = 0.0;
 }
 
-Vec3f::Vec3f(double ix, double iy, double iz)
+Vec3d::Vec3d(double ix, double iy, double iz)
 {
     this->x = ix;
     this->y = iy;
